requirements.txt
Django==2.1.4
django-rest-auth==0.9.3
djangorestframework==3.9.0
Pygments==2.3.0
pytz==2018.7
six==1.11.0
